Fix soundness issue on drop
Turns out that we can't just use `Box` to deallocate nodes because calls
to `dealloc` have to use layouts with the same *alignment* as the
original allocation as well.

In addition to introducing the new `dealloc_aligned` function, this
commit also removes any `Box` construction for convenience in the body
of `NodeHandle::do_drop`.