fix more fully bug from bf573798
We actually need to set parent & index in parent for *many* children
when we insert a new key+child. The indexes are strictly required in
order to account for a shift in the positions of later children, but
it's also worthwhile to opportunistically overwrite child pointers if
they might've changed.