implement another insertion case
This implements the "insert two things to the right" case for insertion,
by extending the existing `insert_rhs` implementation to support an
optional second item.

There's also a whole host of bugs in `insert_rhs` and the various
`do_upward_step` functions that are fixed here as well, and undoubtedly
many more still left to find.

This commit marks a milestone of now being able to *actually* start
fuzzing.