fix mishandling of existing edge cases, +refactor
this is a somewhat broad rework of the existing way that
`do_insert_no_join` is implemented, dividing it into two additional
functions to handle the "split node" cases.

as part of this, many more edge cases are handled now -- all the ones we
currently know of. The new comments introduced in this commit are
expansive; more explaining can be found there.