Fully implement slice references
Lots of changes here, notably adding `const M: usize` parameters in a
lot of places to allow us to actually use the type system in
implementing `SliceRefStore`.

There are, of course, plenty of bugs left in the implementation (some
already discovered in uncommitted tests). Fixing those is left for a
series of follow-up commits.