improve the compiler's knowledge around `ZeroU8`
Honestly not sure whether this has any practical benefits, but the idea
is to ensure that all accesses for `ZeroU8` go through `get`, which
proves to the compiler that the inner value is zero.

It *probably* allows slightly more freedom in accessing the height
value because it's known to be zero. Or maybe not, who knows. Now the
type is at least a little more sound.