Drain pt 2: implement `Drain::{next,next_back}`
This commit also adds the `DrainMut` borrow type for
`{Node,Slice}Handle`s, to allow access to nodes in a way that doesn't
need to uphold the initialization guarantees of slice values and
`RefId`s (i.e., `Leaf.vals` and `Leaf.refs`). Hopefully the
documentation contained there should be enough to understand it as is.