Use trait objects to iter/build `Cursor`s
This is part 2 of a 2-part change to improve codegen size by moving some
more pieces to dyn functions instead of generics.

Doing this for `Cursor`s requires jumping through a few hoops, notably
with the addition of the `CursorBuilder` struct, but also the addition
of a new `'c` lifetime to all of the relevant types and functions in
order to handle it properly.